Moved UserServicer to own file
